pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edgepilot"
version = "0.1.0"
description = "EdgePilot — Your AI Copilot for System Health, Scheduling, and Optimization"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="EdgePilot Team"}]
dependencies = [
  "psutil>=5.9",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "PyYAML>=6.0",
  "sqlalchemy>=2.0",
  "alembic>=1.13",       # optional, here for future migrations
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "jinja2>=3.1",
  "streamlit>=1.37",
  "typer[all]>=0.12",
  "rich>=13.7",
  "python-dateutil>=2.9",
  "mcp>=1.1.1",          # Official Model Context Protocol Python SDK
  "orjson>=3.10",
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "requests>=2.32",
  "aiofiles>=24.1",
  "packaging>=24.1",
]

[project.optional-dependencies]
providers = [
  "anthropic>=0.40 ; platform_system != 'Linux' or platform_machine != 'aarch64'",
  "google-generativeai>=0.7",
]

[project.scripts]
edgepilot = "edgepilot.cli:app"

[tool.pytest.ini_options]
addopts = "-q"
